+++ test/fixtures/cookbooks/dnsmasq-local_test/recipes/default.rb
@@ -1,14 +1,17 @@
 # Encoding: UTF-8
 
-# Fake out the Docker build and its immutable resolv.conf.
-directory '/var/lib/resolvconf'
-file '/var/lib/resolvconf/linkified'
+execute 'apt-get update'
 
-include_recipe 'apt'
-%w(apt-utils resolvconf).each { |p| package p }
+if %w(docker lxc).include?(node['virtualization']['system'])
+  # Fake out the Docker build and its immutable resolv.conf.
+  directory '/var/lib/resolvconf'
+  file '/var/lib/resolvconf/linkified'
 
-service 'resolvconf' do
-  action [:enable, :start]
+  %w(apt-utils resolvconf).each { |p| package p }
+
+  service 'resolvconf' do
+    action [:enable, :start]
+  end
 end
 
 include_recipe 'dnsmasq-local'